+++ app/src/routes/projects/[projectId]/export/download/+server.ts
@@ -1,4 +1,4 @@
-import { buildExportResponse } from './response';
+import { buildExportResponse } from '../response';
 import type { RequestHandler } from './$types';
 
 export const GET: RequestHandler = ({ params }) => buildExportResponse(params.projectId);